fixed foreign key problem and data length not enough

+++ database/mysql/caisi/updates/patch-2008-02-29.sql
@@ -1,4 +1,4 @@
-alter table program add column facility_id bigint(22) not null after program_id;
+alter table program add column facility_id int not null after program_id;
 alter table program add index (facility_id);
 update program,room set program.facility_id=room.facility_id where program.program_id=room.program_id;
 update program set facility_id=1 where facility_id=0;
